adressing naming convention, imports, module location
